refactor : controller in out 분리
