+++ organization-service/src/main/java/com/sparta/moim/organization/presentation/controller/in/OrganizationApplicationController.java
@@ -1,4 +1,4 @@
-package com.sparta.moim.organization.presentation.controller;
+package com.sparta.moim.organization.presentation.controller.in;
 
 
 import com.sparta.moim.common.response.ApiResponseData;
@@ -11,7 +11,6 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
